abstract categories view out of blocks
