+++ src/block_manager/view/BlocksView.js
@@ -1,22 +1,14 @@
-var Backbone = require('backbone');
 var BlockView = require('./BlockView');
-var CategoryView = require('domain_abstract/view/CategoryView');
+var CategorizableView = require('domain_abstract/view/CategorizableView');
 
-module.exports = Backbone.View.extend({
+module.exports = CategorizableView.extend({
+
+  categorizableView: BlockView,
+  categorizableType: 'block',
 
   initialize(opts, config) {
+    CategorizableView.prototype.initialize.apply(this, arguments);
     _.bindAll(this, 'getSorter', 'onDrag', 'onDrop');
-    this.config = config || {};
-    this.categories = opts.categories || '';
-    this.renderedCategories = [];
-    var ppfx = this.config.pStylePrefix || '';
-    this.ppfx = ppfx;
-    this.noCatClass = `${ppfx}blocks-no-cat`;
-    this.blockContClass = `${ppfx}blocks-c`;
-    this.catsClass = `${ppfx}block-categories`;
-    const coll = this.collection;
-    this.listenTo(coll, 'add', this.addTo);
-    this.listenTo(coll, 'reset', this.render);
     this.em = this.config.em;
     this.tac = 'test-tac';
     this.grabbingCls = this.ppfx + 'grabbing';
@@ -86,104 +78,6 @@ module.exports = Backbone.View.extend({
 
       em.trigger('block:drag:stop', model);
     }
-  },
-
-  /**
-   * Add new model to the collection
-   * @param {Model} model
-   * @private
-   * */
-  addTo(model) {
-    this.add(model);
-  },
-
-  /**
-   * Render new model inside the view
-   * @param {Model} model
-   * @param {Object} fragment Fragment collection
-   * @private
-   * */
-  add(model, fragment) {
-    var frag = fragment || null;
-    var view = new BlockView({
-      model,
-      attributes: model.get('attributes'),
-    }, this.config);
-    var rendered = view.render().el;
-    var category = model.get('category');
-
-    // Check for categories
-    if (category && this.categories) {
-      if (typeof category == 'string') {
-        category = {
-          id: category,
-          label: category,
-          type: 'block'
-        };
-      }
-
-      var catModel = this.categories.add(category);
-      var catId = catModel.get('id');
-      var catView = this.renderedCategories[catId];
-      var categories = this.getCategoriesEl();
-      model.set('category', catModel);
-
-      if (!catView && categories) {
-        catView = new CategoryView({
-          model: catModel,
-          contentType: 'block'
-        }, this.config).render();
-        this.renderedCategories[catId] = catView;
-        categories.appendChild(catView.el);
-      }
-
-      catView && catView.append(rendered);
-      return;
-    }
-
-    if(frag)
-      frag.appendChild(rendered);
-    else
-      this.append(rendered);
-  },
-
-  getCategoriesEl() {
-    if (!this.catsEl) {
-      this.catsEl = this.el.querySelector(`.${this.catsClass}`);
-    }
-
-    return this.catsEl;
-  },
-
-  getBlocksEl() {
-    if (!this.blocksEl) {
-      this.blocksEl = this.el.querySelector(`.${this.noCatClass} .${this.blockContClass}`);
-    }
-
-    return this.blocksEl;
-  },
-
-  append(el) {
-    let blocks = this.getBlocksEl();
-    blocks && blocks.appendChild(el);
-  },
-
-  render() {
-    const frag = document.createDocumentFragment();
-    this.catsEl = null;
-    this.blocksEl = null;
-    this.renderedCategories = [];
-    this.el.innerHTML = `
-      <div class="${this.catsClass}"></div>
-      <div class="${this.noCatClass}">
-        <div class="${this.blockContClass}"></div>
-      </div>
-    `;
-
-    this.collection.each(model => this.add(model, frag));
-    this.append(frag);
-    this.$el.addClass(this.blockContClass + 's')
-    return this;
   },
 
 });
