+++ src/domain_abstract/model/Categorizable.js
@@ -1,0 +1,28 @@
+var Backbone = require('backbone');
+var Category = require('./Category');
+
+module.exports = Backbone.Model.extend({
+
+  defaults: {
+    label: '',
+    content: '',
+    category: '',
+    //categorizableType: '',
+    attributes: {},
+  },
+
+  initialize(opts = {}) {
+    let category = this.get('category');
+
+    if (category) {
+      if (typeof category == 'string') {
+        var catObj = new Category({
+          id: category,
+          label: category,
+          type: 'block',
+        });
+      }
+    }
+  },
+
+});
